Moved footer from containers to components
