--- src/containers/app_footer/index.vue
@@ -1,16 +1,0 @@
-
-<template>
-  	<AppFooter/>
-</template>
-
-<!-- // // // //  -->
-
-<script>
-import AppFooter from './components/layout.vue'
-
-export default {
-  components: {
-    AppFooter
-  }
-}
-</script>
